Implemented ProgrammaticMaterialSwitch
I've implemented a new view that allows to programmatically set the checked / unchecked state of MaterialSwitch-es without calling any onCheckedChangeListener.

Ideally, this will be useful for us in the future to restore the state of the switches after the user has clicked them if exceptions occur (for example if the mod could not be applied by the CoreRootService).

Currently the only new case in which the state of a switch is reset is if the user cancels the Dialog of the Suggested Mod "Force enable Call Screen".